backend_python/requirements.txt
flask
mysql-connector-python
pandas
scikit-learn
requests
numpy
